[DP-232] fix: remove collapsible tab scroll
