+++ src/feature/data/components/sections/filter/FilterCardContent.tsx
@@ -9,6 +9,7 @@ interface FilterCardContentProps {
   max: number;
   onValueChange: (val: number[]) => void;
   disabled?: boolean;
+  gap?: number;
 }
 
 export default function FilterCardContent({
@@ -17,19 +18,18 @@ export default function FilterCardContent({
   value,
   max,
   onValueChange,
+  gap = 12,
   disabled,
 }: FilterCardContentProps) {
   return (
-    <div className="flex flex-col h-full justify-center items-center text-center gap-12 p-12">
+    <div
+      className="flex flex-col h-full justify-center items-center text-center p-12"
+      style={{ gap: gap }}
+    >
       <h2 className="h1 text-black">{formatDataSize(value[0])}</h2>
       <p className="body-sm text-gray-600">원하는 용량을 선택하세요</p>
 
-      <Slider 
-        value={value} 
-        onValueChange={onValueChange} 
-        max={max} 
-        disabled={disabled}
-      />
+      <Slider value={value} onValueChange={onValueChange} max={max} disabled={disabled} />
 
       <ButtonComponent
         variant="nonoutline"
